"Added new service and repository classes for User, Feedback, Trip, and Driver entities
Introduced new service and repository classes for the Feedback, User, Trip, and Driver entities to enhance the functionality of the travel services. With the Feedback service, users now have the ability to rate and review trips, while the Trip service provides details about the trips made. The User service was also updated to efficiently manage user-related information like getting user details by username in addition to the existing id-based access. Similarly, a Driver service is implemented that provides functionalities to manage data related the drivers."